Refactor App component to use separate tabs for Dashboard, Inventory, and Transactions; add new components for each tab.
